+++ mayan/apps/common/apps.py
@@ -18,7 +18,7 @@ from .links import (
     link_about, link_knowledge_base, link_license, link_setup, link_support,
     link_tools
 )
-from .menus import menu_about, menu_topbar, menu_user
+from .menus import menu_system, menu_topbar, menu_user
 from .settings import setting_home_view
 from .signals import signal_pre_initial_setup, signal_pre_upgrade
 
@@ -146,7 +146,7 @@ class CommonApp(MayanAppConfig):
             template_name='appearance/menus/topbar.html'
         )
 
-        menu_about.bind_links(
+        menu_system.bind_links(
             links=(
                 link_tools, link_setup, link_knowledge_base, link_support,
                 link_about, link_license
@@ -154,7 +154,7 @@ class CommonApp(MayanAppConfig):
         )
 
         menu_topbar.bind_links(
-            links=(menu_about, menu_user),
+            links=(menu_system, menu_user),
             position=10
         )