+++ mayan/apps/common/menus.py
@@ -2,11 +2,8 @@ from django.utils.translation import gettext_lazy as _
 
 from mayan.apps.navigation.classes import Menu
 
-from .icons import icon_menu_about, icon_menu_user
+from .icons import icon_menu_system, icon_menu_user
 
-menu_about = Menu(
-    icon=icon_menu_about, label=_(message='System'), name='about'
-)
 menu_facet = Menu(label=_(message='Facet'), name='facet')
 menu_list_facet = Menu(label=_(message='Facet'), name='list facet')
 menu_main = Menu(name='main')
@@ -15,6 +12,9 @@ menu_object = Menu(label=_(message='Actions'), name='object')
 menu_related = Menu(label=_(message='Related'), name='related')
 menu_secondary = Menu(label=_(message='Secondary'), name='secondary')
 menu_setup = Menu(name='setup')
+menu_system = Menu(
+    icon=icon_menu_system, label=_(message='System'), name='about'
+)
 menu_return = Menu(label=_(message='Return'), name='return')
 menu_tools = Menu(name='tools')
 menu_topbar = Menu(name='topbar')